refactor: agno/tools/component_tool.py -> agno/tools/component.py
Single-word module name matching the folder's convention; the class name
and the documented import (from agno.tools import ComponentTool) are
unchanged. Free while unreleased.
