+++ libs/agno/agno/tools/__init__.py
@@ -1,4 +1,4 @@
-from agno.tools.component_tool import ComponentTool
+from agno.tools.component import ComponentTool
 from agno.tools.decorator import tool
 from agno.tools.function import Function, FunctionCall
 from agno.tools.toolkit import Toolkit
