+++ libs/agno/agno/workflow/workflow.py
@@ -29,7 +29,7 @@ if TYPE_CHECKING:
     from fastapi import WebSocket
 
     from agno.os.managers import WebSocketHandler
-    from agno.tools.component_tool import ComponentTool
+    from agno.tools.component import ComponentTool
 else:
     # fastapi only ships with the "os" extra. Binding WebSocket loosely keeps
     # the websocket annotations resolvable by get_type_hints() -- and with
@@ -737,14 +737,14 @@ class Workflow:
     def as_tool(self, name: Optional[str] = None, description: Optional[str] = None) -> "ComponentTool":
         """Publish this workflow as a tool with its own model-facing name and description.
 
-        Returns a declarative :class:`~agno.tools.component_tool.ComponentTool` marker
+        Returns a declarative :class:`~agno.tools.component.ComponentTool` marker
         for surfaces that turn components into tools -- today the AgentOS MCP server:
         ``MCPConfig(tools=[brief.as_tool(name="ask_brief", description=...)])``. Both
         overrides are optional; the tool name must be a valid tool identifier (start
         with a letter or underscore, then letters/digits/hyphens/underscores). The
         workflow id remains the continue_run handle and the scope segment.
         """
-        from agno.tools.component_tool import ComponentTool
+        from agno.tools.component import ComponentTool
 
         return ComponentTool(component=self, name=name, description=description)
